- Update Polly dependency to latest. - Consolidate solution and modernise build. - Add SourceLink support. - Add NetStandard 2.1 target (for .NET Core3.0 consumption) - Add test runs in netcoreapp3.0; .NET Framework 4.6.1; and .NET Framework 4.7.2 - Update FluentAssertions and xUnit dependencies - Remove duplicated test cases
