+++ GitVersionConfig.yaml
@@ -1,1 +1,1 @@
-next-version: 3.0.0
+next-version: 3.0.1
