+++ appveyor.yml
@@ -1,4 +1,4 @@
-os: Visual Studio 2017
+image: Visual Studio 2019
 
 # Build script
 build_script:
@@ -9,6 +9,7 @@ test: off
 
 artifacts:
   - path: artifacts\nuget-package\*.nupkg
+  - path: artifacts\nuget-package\*.snupkg
 
 environment:
   # Skip dotnet package caching on build servers
