--- src/Polly.Caching.Distributed.NetStandard20.Specs/Properties/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-﻿using System.Reflection;
-using Xunit;
-
-[assembly: AssemblyTitle("Polly.Caching.Distributed.NetStandard20.Specs")]
-[assembly: CollectionBehavior(DisableTestParallelization = true)]
