--- src/Polly.Caching.Distributed.SharedSpecs/Polly.Caching.Distributed.SharedSpecs.projitems
@@ -1,27 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>8d537c37-a4a2-4be1-84bc-34f1afb51ca8</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>Polly.Caching.Distributed.Specs</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\ByteArraySerializer.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\MemoryDistributedCachePolicyFactory.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\CacheRoundTripSpecsAsyncBase.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\CacheRoundTripSpecsBase.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\CacheRoundTripSpecsSyncBase.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\CacheRoundTripSpecs_NetStandardIDistributedCacheProvider_Async_ByteArray.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\CacheRoundTripSpecs_NetStandardIDistributedCacheProvider_Async_String.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\CacheRoundTripSpecs_NetStandardIDistributedCacheProvider_Sync_ByteArray.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\CacheRoundTripSpecs_NetStandardIDistributedCacheProvider_Sync_String.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Integration\ICachePolicyFactory.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Unit\CacheProviderHelperTests.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Unit\NetStandardIDistributedCacheByteArrayProviderSpecs.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Unit\NetStandardIDistributedCacheStringProviderSpecs.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Unit\TtlStrategyHelperTests.cs" />
-  </ItemGroup>
-</Project>
