+++ src/Polly.Caching.Distributed.Specs/Integration/ByteArraySerializer.cs
