+++ src/Polly.Caching.Distributed.Specs/Integration/CacheRoundTripspecsAsyncBase.cs
@@ -28,7 +28,7 @@ namespace Polly.Caching.Distributed.Specs.Integration
                     underlyingDelegateExecuteCount++;
                     return Task.FromResult(testValue);
                 }, new Context(OperationKey)))
-                .ShouldBeEquivalentTo(testValue);
+                .Should().BeEquivalentTo(testValue);
 
             // Assert - should have executed underlying delegate
             underlyingDelegateExecuteCount.Should().Be(1);
@@ -36,7 +36,7 @@ namespace Polly.Caching.Distributed.Specs.Integration
             // Assert - should be in cache
             (bool cacheHit2, TResult fromCache2) = await cacheProvider.TryGetAsync(OperationKey, CancellationToken.None, false);
             cacheHit2.Should().BeTrue();
-            fromCache2.ShouldBeEquivalentTo(testValue);
+            fromCache2.Should().BeEquivalentTo(testValue);
 
             // Act - should execute underlying delegate and place in cache
             (await cache.ExecuteAsync(ctx =>
@@ -44,7 +44,7 @@ namespace Polly.Caching.Distributed.Specs.Integration
                     underlyingDelegateExecuteCount++;
                     throw new Exception("Cache should be used so this should not get invoked.");
                 }, new Context(OperationKey)))
-                .ShouldBeEquivalentTo(testValue);
+                .Should().BeEquivalentTo(testValue);
             underlyingDelegateExecuteCount.Should().Be(1);
         }
     }
