+++ src/Polly.Caching.Distributed.Specs/Integration/ICachePolicyFactory.cs
