--- src/Polly.Caching.Distributed.nuspec
@@ -1,56 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package>
-  <metadata>
-    <owners>App vNext</owners>
-    <authors>App vNext</authors>
-    <description>
-      Polly.Caching.Distributed is a plug-in for the .NET OSS resilience library Polly, supporting Microsoft.Extensions.Caching.Distributed.IDistributedCache as a provider for Polly's CachePolicy.
-    </description>
-    <language>en-US</language>
-    <license type="expression">BSD-3-Clause</license>
-    <iconUrl>https://raw.github.com/App-vNext/Polly/master/Polly.png</iconUrl>
-    <projectUrl>https://github.com/App-vNext/Polly.Caching.IDistributedCache</projectUrl>
-    <tags>Polly Cache Caching Cache-aside</tags>
-    <copyright>Copyright © 2019, App vNext</copyright>
-    <releaseNotes>
-     3.0.0
-     ---------------------
-     - Allow caching of default(TResult)
-     - Compatible with Polly &gt;= v7
-
-     2.0.1
-     ---------------------
-     - No functional changes
-     - Indicate compatibility with Polly &lt; v7
-
-     2.0.0
-     ---------------------
-     - Provide a single signed package only.
-     - Update Polly to V6.0.1.
-     - Support .net Standard 2.0.
-     - Rename package, and change namespaces, from Polly.Caching.IDistributedCache to Polly.Caching.Distributed, to avoid clashes.
-
-     1.0-RC
-     ---------------------
-
-     - Allows Polly CachePolicy to be used with Microsoft.Extensions.Caching.Distributed.IDistributedCache implementations.
-     - Release candidate
-
-   </releaseNotes>
-    <dependencies>
-      <group targetFramework="netstandard1.1">
-        <dependency id="NETStandard.Library" version="1.6.1" />
-        <dependency id="Polly" version="7.0.0" />
-        <dependency id="Microsoft.Extensions.Caching.Abstractions" version="1.1.2" />
-        <dependency id="System.ValueTuple" version="4.5.0" />
-      </group>
-      <group targetFramework="netstandard2.0">
-        <dependency id="Polly" version="7.0.0" />
-        <dependency id="Microsoft.Extensions.Caching.Abstractions" version="2.0.2" />
-      </group>
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="lib\**\*.*" target="lib" exclude="**\*.unsigned" />
-  </files>
-</package>
