+++ src/Polly.Caching.Distributed.sln
@@ -1,52 +1,26 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.16
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29403.142
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{79FE9DBE-1155-4F78-A41F-FE95DEFA19DD}"
-	ProjectSection(SolutionItems) = preProject
-		GlobalAssemblyInfo.cs = GlobalAssemblyInfo.cs
-		Polly.nuspec = Polly.nuspec
-	EndProjectSection
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.Distributed", "Polly.Caching.Distributed\Polly.Caching.Distributed.csproj", "{78D2B68F-131C-43DD-820B-D761C30AB2CD}"
 EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Polly.Caching.Distributed.Shared", "Polly.Caching.Distributed.Shared\Polly.Caching.Distributed.Shared.shproj", "{23FA87F9-C77D-4C67-A0B0-2901DE51B3FF}"
-EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Polly.Caching.Distributed.SharedSpecs", "Polly.Caching.Distributed.SharedSpecs\Polly.Caching.Distributed.SharedSpecs.shproj", "{8D537C37-A4A2-4BE1-84BC-34F1AFB51CA8}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.Distributed.NetStandard11", "Polly.Caching.Distributed.NetStandard11\Polly.Caching.Distributed.NetStandard11.csproj", "{44926F85-3309-4D4E-BD6E-AAEF076B9D6A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.Distributed.NetStandard11.Specs", "Polly.Caching.Distributed.NetStandard11.Specs\Polly.Caching.Distributed.NetStandard11.Specs.csproj", "{E5EB8F22-174A-4D2B-A0A1-4C2F82883F8E}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.Distributed.NetStandard20", "Polly.Caching.Distributed.NetStandard20\Polly.Caching.Distributed.NetStandard20.csproj", "{67FD1222-E88B-4472-A84C-66752020126E}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.Distributed.NetStandard20.Specs", "Polly.Caching.Distributed.NetStandard20.Specs\Polly.Caching.Distributed.NetStandard20.Specs.csproj", "{1A92408B-54BF-40C9-A2A0-3BC49B174A07}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.Distributed.Specs", "Polly.Caching.Distributed.Specs\Polly.Caching.Distributed.Specs.csproj", "{732E300F-20CD-4F1F-A07B-2A5FA17F9789}"
 EndProject
 Global
-	GlobalSection(SharedMSBuildProjectFiles) = preSolution
-		Polly.Caching.Distributed.Shared\Polly.Caching.Distributed.Shared.projitems*{23fa87f9-c77d-4c67-a0b0-2901de51b3ff}*SharedItemsImports = 13
-		Polly.Caching.Distributed.SharedSpecs\Polly.Caching.Distributed.SharedSpecs.projitems*{8d537c37-a4a2-4be1-84bc-34f1afb51ca8}*SharedItemsImports = 13
-	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{44926F85-3309-4D4E-BD6E-AAEF076B9D6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{44926F85-3309-4D4E-BD6E-AAEF076B9D6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{44926F85-3309-4D4E-BD6E-AAEF076B9D6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{44926F85-3309-4D4E-BD6E-AAEF076B9D6A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E5EB8F22-174A-4D2B-A0A1-4C2F82883F8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E5EB8F22-174A-4D2B-A0A1-4C2F82883F8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E5EB8F22-174A-4D2B-A0A1-4C2F82883F8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E5EB8F22-174A-4D2B-A0A1-4C2F82883F8E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{67FD1222-E88B-4472-A84C-66752020126E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{67FD1222-E88B-4472-A84C-66752020126E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{67FD1222-E88B-4472-A84C-66752020126E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{67FD1222-E88B-4472-A84C-66752020126E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1A92408B-54BF-40C9-A2A0-3BC49B174A07}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1A92408B-54BF-40C9-A2A0-3BC49B174A07}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1A92408B-54BF-40C9-A2A0-3BC49B174A07}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1A92408B-54BF-40C9-A2A0-3BC49B174A07}.Release|Any CPU.Build.0 = Release|Any CPU
+		{78D2B68F-131C-43DD-820B-D761C30AB2CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{78D2B68F-131C-43DD-820B-D761C30AB2CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{78D2B68F-131C-43DD-820B-D761C30AB2CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{78D2B68F-131C-43DD-820B-D761C30AB2CD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{732E300F-20CD-4F1F-A07B-2A5FA17F9789}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{732E300F-20CD-4F1F-A07B-2A5FA17F9789}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{732E300F-20CD-4F1F-A07B-2A5FA17F9789}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{732E300F-20CD-4F1F-A07B-2A5FA17F9789}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
