+++ src/Polly.Caching.Distributed/NetStandardIDistributedCacheByteArrayProvider.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
 
 namespace Polly.Caching.Distributed
